requirements.txt
asgiref==3.11.0
asttokens==3.0.1
black==25.12.0
boto3==1.42.44
botocore==1.42.44
click==8.3.1
decorator==5.2.1
Django==6.0.1
django-environ==0.12.0
django-storages==1.14.6
executing==2.2.1
gunicorn==24.1.1
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.1.0
matplotlib-inline==0.2.1
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.5
pathspec==1.0.1
pexpect==4.9.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psycopg==3.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytokens==0.3.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.6.3
wcwidth==0.2.14
whitenoise==6.11.0
